gitflow-feature-stash: Comestique
Quelques améliorations de code, quelques lignes de documentation et des questionnements
