+++ src/Enums/ApiModeEnum.php
@@ -1,0 +1,25 @@
+<?php
+/* This file is part of NextDom.
+ *
+ * NextDom is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * NextDom is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with NextDom. If not, see <http://www.gnu.org/licenses/>.
+ */
+namespace NextDom\Enums;
+
+class ApiModeEnum extends Enum
+{
+    const API_ENABLE     = 'enable';
+    const API_DISABLE    = 'disable';
+    const API_WHITEIP    = 'whiteip';
+    const API_LOCALHOST  = 'cronDaily';
+}
